+++ src/app/[variants]/(main)/knowledge/routes/KnowledgeBaseDetail/menu/Menu.tsx
@@ -3,14 +3,14 @@
 import { memo } from 'react';
 import { Flexbox } from 'react-layout-kit';
 
+import CategoryMenu from './CategoryMenu';
 import Head from './Head';
-import MenuItems from './MenuItems';
 
 const Menu = memo<{ id: string }>(({ id }) => {
   return (
     <Flexbox gap={16} height={'100%'} paddingInline={12} style={{ paddingTop: 12 }}>
       <Head id={id} />
-      <MenuItems />
+      <CategoryMenu />
     </Flexbox>
   );
 });
