+++ src/features/KnowledgeManager/ChunkDrawer/SimilaritySearchList/Item.tsx
