+++ src/features/KnowledgeManager/DocumentExplorer/NoteEditorModal.tsx
