+++ src/features/KnowledgeManager/DocumentExplorer/NoteEditorPanel.tsx
