+++ src/features/KnowledgeManager/DocumentExplorer/NoteEmptyStatus.tsx
