+++ src/features/KnowledgeManager/FileExplorer/FileSkeleton.tsx
