+++ src/features/KnowledgeManager/FileExplorer/MasonryFileItem/DefaultFileItem.tsx
