+++ src/features/KnowledgeManager/FileExplorer/MasonryFileItem/ImageFileItem.tsx
