+++ src/features/KnowledgeManager/FileExplorer/MasonryFileItem/MarkdownFileItem.tsx
