+++ src/features/KnowledgeManager/FileExplorer/MasonryFileItem/MasonryItemWrapper.tsx
