+++ src/features/KnowledgeManager/FileExplorer/MasonryFileItem/NoteFileItem.tsx
