+++ src/features/KnowledgeManager/FileExplorer/MasonryFileItem/index.tsx
