+++ src/features/KnowledgeManager/FileExplorer/MasonrySkeleton.tsx
