+++ src/features/KnowledgeManager/FileExplorer/ToolBar/ViewSwitcher.tsx
