+++ src/features/KnowledgeManager/Header/FilesSearchBar.tsx
