+++ src/features/KnowledgeManager/Header/NewNoteButton.tsx
