+++ src/features/KnowledgeManager/Header/TogglePanelButton.tsx
