+++ src/features/KnowledgeManager/UploadDock/Item.tsx
