+++ src/features/KnowledgeManager/UploadDock/index.tsx
